+++ MT-NETCore-Data/Entities/ProjectUser.cs
@@ -6,9 +6,9 @@ namespace MT_NetCore_Data.Entities
     //Join Table for many to many relationship between Projects and Users.
     public class ProjectUser 
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
-        public int ProjectId { get; set; }
+        public long ProjectId { get; set; }
         public Project Project { get; set; }
         public Role UserRole { get; set; }
